Cleanup certificate uploading instructions.
Removed the suggested home folder upload location, leaving the following bullet point's suggested paths in place, as these better lines up with the example vHost provided further down the article.

Reordered the notes and bullet points to match the order in which those steps would need to be taken by the reader.

Updated the certificate copying note to suggest a name including the proper file extension. Removed references to uploading the CSR itself, which is not actually used in the vHost configuration, and can be discarded once the certificate is obtained.