The confusion of making a https CDN out of http only origin URL
I have made changes that shows that in Rackspace CDN you can only make http CDN endpoint and not https with s3 static site. 

If it is in fact possible to create https CDN endpoint for s3 static site then please show how. (It is possible to do so in cloudfront and cloudflare with s3 static site)