Fix broken link
On line 22, there is a broken link on the page. It looks to be correct on the format, but it's not creating the link on the page. 

I had to add an extra line to submit the change request as I'm not sure what needs to change on line 22 to display the correct link.